+++ src/common/js/configure.js
@@ -6,6 +6,28 @@ import _ from 'lodash';
 
 CookieManager._conf = {};
 
+const _checkCategories = function( name ){
+    const val = CookieManager._conf[name];
+    let errs = 0;
+    let newvals = [];
+    let newcats = {};
+    if( !_.isArray( val )){
+        console.warn( 'pwix:cookie-manager expects', name, 'be an array, found', val, 'reset to', CookieManager._defaults[name] );
+    } else {
+        val.every(( it ) => {
+            if( !it.id || !_.isString( it.id )){
+                console.warn( 'pwix:cookie-manager expects eacnh category have its own identifier, found', it.id, 'removing it' );
+            } else {
+                newvals.push( it );
+                newcats[it.id] = it;
+            }
+            return true;
+        });
+        CookieManager._conf[name] = newvals;
+        _.merge( CookieManager.Cdyn.Category, CookieManager.C.Category, newcats );
+    }
+};
+
 const _checkInteger = function( name ){
     const val = CookieManager._conf[name];
     const ival = parseInt( val );
@@ -16,6 +38,7 @@ const _checkInteger = function( name ){
 };
 
 CookieManager._defaults = {
+    categories: [],
     consentLifetime: 31536000000,   // one year
     verbosity: CookieManager.C.Verbose.NONE
 };
@@ -32,6 +55,7 @@ CookieManager.configure = function( o ){
     if( o && _.isObject( o )){
         _.merge( CookieManager._conf, CookieManager._defaults, o );
         // make sure we have integer where required
+        _checkCategories( 'categories' );
         _checkInteger( 'consentLifetime' );
         // be verbose if asked for
         if( CookieManager._conf.verbosity & CookieManager.C.Verbose.CONFIGURE ){